==> notebooks/01_bigquery_analysis.ipynb <==
from google.cloud import bigquery
project_id="students-group2"
client = bigquery.Client(project=project_id)

tables=client.list_tables("students-group2.fatma_aziz")
for i in tables:
    print(i.table_id) 

# les noms des colonnes de la table movies 
movies = client.get_table("students-group2.fatma_aziz.movies")
print([schema_field.name for schema_field in movies.schema])

# le nom des colonnes de la table ratings 
ratings = client.get_table("students-group2.fatma_aziz.ratings")
print([schema_field.name for schema_field in ratings.schema])

# Dataframes extraction
df_movies = client.query("SELECT movieId, title, genres FROM `students-group2.fatma_aziz.movies`").to_dataframe()
df_ratings = client.query("SELECT userId, movieId, rating, timestamp FROM `students-group2.fatma_aziz.ratings`").to_dataframe()

df_movies

df_movies.info()

df_ratings

df_ratings.info()

print("le nombre de films total:",len(df_movies['movieId'].value_counts()))

print("le nombre de films ayant des ratings:",len(df_ratings['movieId'].value_counts()))

print("le nombre d'utilisateurs:",len(df_ratings['userId'].value_counts()))

print("les valeurs possibles de rating:", df_ratings['rating'].value_counts())

print("les genres de films:", df_movies['genres'].value_counts())

# Essayer des requetes BigQuery 
avg_rating_per_movie = client.query("""
    SELECT movieId, AVG(rating) AS avg_rating, COUNT(*) AS n_ratings
    FROM `students-group2.fatma_aziz.ratings`
    GROUP BY movieId
    ORDER BY avg_rating DESC
""").to_dataframe()

avg_rating_per_movie

movie_avg_ratings = client.query("""
    SELECT m.movieId, m.title, AVG(r.rating) AS avg_rating
    FROM `students-group2.fatma_aziz.ratings` r
    JOIN `students-group2.fatma_aziz.movies` m
    ON r.movieId = m.movieId
    GROUP BY m.movieId, m.title
    ORDER BY avg_rating DESC
""").to_dataframe()

movie_avg_ratings

# Nombre de rating par utilisateur
ratings_per_user = client.query("""
    SELECT userId, COUNT(*) AS n_ratings
    FROM `students-group2.fatma_aziz.ratings`
    GROUP BY userId
    ORDER BY n_ratings DESC
""").to_dataframe()


ratings_per_user

df_movies_no_ratings = client.query("""
SELECT 
  m.movieId,
  m.title,
  m.genres
FROM `students-group2.fatma_aziz.movies` m
LEFT JOIN `students-group2.fatma_aziz.ratings` r
  ON m.movieId = r.movieId
WHERE r.movieId IS NULL
"""
).to_dataframe()

print("Nombre de films sans ratings :", df_movies_no_ratings.shape[0])
df_movies_no_ratings.head()

# “We identify cold-start items (movies without ratings) using BigQuery.”

df_number_movie_per_genre = client.query("""
SELECT 
  genres,
  COUNT(*) AS number_of_movies
FROM `students-group2.fatma_aziz.movies`
GROUP BY genres
ORDER BY number_of_movies DESC
""").to_dataframe()

print("Number of movies per genres")
df_number_movie_per_genre.head(10)
