==> notebooks/03_preprocessing.ipynb <==
from google.cloud import bigquery
project_id="students-group2"
client = bigquery.Client(project=project_id)

# Dataframes 
df_movies = client.query("SELECT * FROM `students-group2.fatma_aziz.movies`").to_dataframe()
df_ratings = client.query("SELECT * FROM `students-group2.fatma_aziz.ratings`").to_dataframe()

# Vérifier s'il y a des doublons 
print("Doublons dans ratings :", df_ratings.duplicated().sum())
print("Doublons dans movies :", df_movies.duplicated().sum())

df_movies['year'] = df_movies['title'].str.extract(r'\((\d{4})\)')[0].astype('Int64')

df_movies['genres_list'] = df_movies['genres'].str.split('|')

df_movies

df_ratings

from sklearn.preprocessing import LabelEncoder

user_encoder = LabelEncoder()
movie_encoder = LabelEncoder()

df_ratings['user_idx'] = user_encoder.fit_transform(df_ratings['userId'])
df_ratings['movie_idx'] = movie_encoder.fit_transform(df_ratings['movieId'])

from sklearn.model_selection import train_test_split

train_df, test_df = train_test_split(df_ratings, test_size=0.2, random_state=42)

# Compter combien de notes chaque utilisateur a fait
user_counts = train_df.groupby('user_idx')['rating'].count()
print(user_counts)

# Compter combien de notes chaque film a reçu
movie_counts = train_df.groupby('movie_idx')['rating'].count()
print(movie_counts)


# Sauvegarder les données et les encoders

import os
from pathlib import Path
os.chdir('/home/jupyter/Fatma_Aziz/movie_recommender')
print(f"📌 Nouveau dossier: {os.getcwd()}")

# Chemin relatif vers la racine du projet
data_path = os.path.join(os.getcwd(), 'data') 

# Créer le dossier data à la racine si pas déjà existant
os.makedirs(data_path, exist_ok=True)

print("✅Le dossier data sera créé ici :", data_path)

# Sauvegarder les données dans un dossier data 
df_ratings.to_csv(os.path.join(data_path, 'df_ratings.csv'), index=False)
train_df.to_csv(os.path.join(data_path, 'train_ratings.csv'), index=False)
test_df.to_csv(os.path.join(data_path, 'test_ratings.csv'), index=False)
df_movies.to_csv(os.path.join(data_path, 'movies.csv'), index=False)

df_ratings.columns

import pickle
# Sauvegarder les encoders
with open(os.path.join(data_path, 'user_encoder.pkl'), 'wb') as f:
    pickle.dump(user_encoder, f)

with open(os.path.join(data_path, 'movie_encoder.pkl'), 'wb') as f:
    pickle.dump(movie_encoder, f)

print("✅ Encoders sauvegardés dans", data_path)

from google.cloud import storage
import pandas as pd
from io import StringIO

BUCKET_NAME = "movie-reco-models-fatma-aziz-students-group2"
DATA_PREFIX = "data/"

def save_df_to_gcs(df: pd.DataFrame, filename: str):
    """
    Sauvegarde un DataFrame pandas sur GCS en CSV.
    """
    client = storage.Client()
    bucket = client.bucket(BUCKET_NAME)
    blob = bucket.blob(DATA_PREFIX + filename)

    csv_buffer = StringIO()
    df.to_csv(csv_buffer, index=False)

    blob.upload_from_string(
        csv_buffer.getvalue(),
        content_type="text/csv"
    )

    print(f"✅ Uploaded {filename} to gs://{BUCKET_NAME}/{DATA_PREFIX}{filename}")


from src.load_save_data import save_data_to_both
# save_df_to_gcs(df_ratings, "df_ratings.csv")
# save_df_to_gcs(train_df, "train_ratings.csv")
# save_df_to_gcs(test_df, "test_ratings.csv")
# save_df_to_gcs(df_movies, "movies.csv")

save_data_to_both(df_ratings, "df_ratings.csv")
save_data_to_both(train_df, "train_ratings.csv")
save_data_to_both(test_df, "test_ratings.csv")
save_data_to_both(df_movies, "movies.csv")
