==> notebooks/04_training.ipynb <==
import os
import pandas as pd
from sklearn.metrics.pairwise import cosine_similarity
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

# Emplacement des données
os.chdir('/home/jupyter/Fatma_Aziz/movie_recommender')
project_root = os.getcwd()  
data_path = os.path.join(project_root, 'data')
os.makedirs(data_path, exist_ok=True)
print("📌 Chemin vers data :", data_path)

# Charger train, test et movies
train_df = pd.read_csv(os.path.join(data_path, 'train_ratings.csv'))
test_df = pd.read_csv(os.path.join(data_path, 'test_ratings.csv'))
df_movies = pd.read_csv(os.path.join(data_path, 'movies.csv'))

# Vérifier
print("Train shape :", train_df.shape)
print("Test shape  :", test_df.shape)
print("Movies shape:", df_movies.shape)


import pickle

with open(os.path.join(data_path, 'user_encoder.pkl'), 'rb') as f:
    user_encoder = pickle.load(f)

with open(os.path.join(data_path, 'movie_encoder.pkl'), 'rb') as f:
    movie_encoder = pickle.load(f)


# filtrage collaboratif 
# user_idx et movie_idx sont déjà créés
user_movie_matrix = train_df.pivot(index='user_idx', columns='movie_idx', values='rating')

# Affichage rapide
print(user_movie_matrix.head())

# Remplacer les NaN par 0 pour calculer la similarité
user_movie_matrix_filled = user_movie_matrix.fillna(0)

# Similarité entre utilisateurs
user_similarity = cosine_similarity(user_movie_matrix_filled)

# Convertir en DataFrame pour plus de clarté
user_similarity_df = pd.DataFrame(user_similarity, 
                                  index=user_movie_matrix.index, 
                                  columns=user_movie_matrix.index)

user_movie_matrix_filled.head(100)

#BASELINE MODEL
# Moyenne des ratings par film
mean_ratings = user_movie_matrix.mean(axis=0)  # moyenne colonne par colonne (chaque film)

# Trier par ordre décroissant
mean_ratings_sorted = mean_ratings.sort_values(ascending=False)

# Afficher les 10 premiers
print(mean_ratings_sorted.head(10))


from surprise import Dataset, Reader, SVD
from surprise.model_selection import train_test_split
from surprise import accuracy
from surprise import Trainset

# Définir le format des données
reader = Reader(rating_scale=(train_df['rating'].min(), train_df['rating'].max()))

# Charger directement le trainset
train_data = Dataset.load_from_df(train_df[['user_idx', 'movie_idx', 'rating']], reader)
trainset = train_data.build_full_trainset()  # Utilise tout train_df

# Charger directement le testset (test_df)
testset = list(zip(test_df['user_idx'], test_df['movie_idx'], test_df['rating']))

# Créer et entraîner le modèle
algo = SVD()
algo.fit(trainset)

# Prédictions sur test set
predictions = algo.test(testset)

# RMSE
rmse = accuracy.rmse(predictions)
print("RMSE sur le test set:", rmse)


# Afficher les 10 premières prédictions et les vraies valeurs
print("user_idx | movie_idx | rating réel | rating prédit")
for pred in predictions[:10]:
    print(f"{pred.uid}        {pred.iid}         {pred.r_ui:.2f}        {pred.est:.2f}")



# Matrice prédite
n_users = train_df['user_idx'].nunique()
n_movies = train_df['movie_idx'].nunique()

pred_matrix = np.zeros((n_users, n_movies))

for uid in range(n_users):
    for iid in range(n_movies):
        pred_matrix[uid, iid] = algo.predict(uid, iid).est
        
pred_df = pd.DataFrame(pred_matrix, 
                       index=[f"User {uid}" for uid in range(n_users)],
                       columns=df_movies['title'].values[:n_movies])  # limiter au nombre de films entraînés

# Afficher les 10 premières lignes et colonnes
print(pred_df.iloc[:10, :10])



## Implémentation manuelle du filtrage collaboratif

# recommandation for users 
def recommand_movie_to_user(user_id, top_n=5):
    # Tous les movie_idx vus par l'utilisateur
    movies_watched = train_df[train_df['user_idx']==user_id]['movie_idx'].tolist()

    # Tous les movie_idx connus par le modèle
    all_movie_idx = list(movie_encoder.transform(movie_encoder.classes_))

    pred_for_user = []

    for movie_idx in all_movie_idx:
        if movie_idx not in movies_watched:
            pred_rating = algo.predict(user_id, movie_idx).est
            pred_for_user.append((movie_idx, pred_rating))

    # Trier par note prédite décroissante
    top_recommendations = sorted(pred_for_user, key=lambda x: x[1], reverse=True)[:5]

    # Afficher
    print("Top 5 recommandations pour user", user_id)
    for movie_idx, rating in top_recommendations:
        movie_id = movie_encoder.inverse_transform([movie_idx])[0]
        title = df_movies[df_movies['movieId']==movie_id]['title'].values[0]
        print(f"{title} - predicted rating: {rating:.2f}")
        
    return sorted(pred_for_user, key=lambda x: x[1], reverse=True)[:top_n]


recommand_movie_to_user(0)

recommand_movie_to_user(50)

## Filtrage collaboratif basé sur les utilisateurs (User-User Collaborative Filtering)

n_users = pred_matrix.shape[0]  
all_user_ids = list(range(n_users))

from sklearn.metrics.pairwise import cosine_similarity
def users_close(user_idx, k=15):
    user_similarity = cosine_similarity(pred_matrix)
    
    user_similarity_df = pd.DataFrame(
        user_similarity,
        index=all_user_ids,
        columns=all_user_ids
    )
    
    closest_users = (
        user_similarity_df[user_idx]
        .sort_values(ascending=False)
        .iloc[1:k+1]   # on enlève l'utilisateur lui-même
    )
    
    return closest_users

users_close(0)

users_close(50)

train_df.columns

movie_encoder.classes_

# 1. Matrice utilisateur × film
user_movie_matrix = train_df.pivot(
    index='user_idx',
    columns='movie_idx',
    values='rating'
)

# 2. Similarité utilisateur-utilisateur
user_similarity = cosine_similarity(user_movie_matrix.fillna(0))
user_similarity_df = pd.DataFrame(
    user_similarity,
    index=user_movie_matrix.index,
    columns=user_movie_matrix.index
)
user_similarity_df

user_id = 0
k = 10

# k plus proches voisins
neighbors = (
    user_similarity_df[user_id]
    .sort_values(ascending=False)
    .iloc[1:k+1]
    .index
)


def avg_neighbor_rating(movie_idx, neighbors):
    ratings = user_movie_matrix.loc[neighbors, movie_idx].dropna()
    return ratings.mean() if len(ratings) > 0 else None


# teste pour l'utilisateur 
user_id = all_user_ids[0] 
neighbors = users_close(user_id).index.tolist()
print("Voisins :", neighbors)


user_id = all_user_ids[0]

recommendations = recommand_movie_to_user(user_id)
neighbors = users_close(user_id).index.tolist()

print(f"\n🎬 Recommandations pour l'utilisateur {user_id} :\n")

for movie_idx, pred_rating in recommendations:
    
    avg_neighbors = avg_neighbor_rating(movie_idx, neighbors)

    movie_id = movie_encoder.inverse_transform([movie_idx])[0]
    title = df_movies.loc[df_movies['movieId'] == movie_id, 'title'].values[0]

    print(f"Film : {title}")
    print(f"  ⭐ Note prédite (SVD)      : {pred_rating:.2f}")
    print(f"  👥 Moyenne des voisins    : {avg_neighbors:.2f}" if avg_neighbors is not None else
          "  👥 Moyenne des voisins    : aucune note")
    print("-" * 50)


## Tester quelques modèles de la bibliothèque Surprise

 # Modèle simple pour comparer les performances
from surprise import BaselineOnly

algo = BaselineOnly()
algo.fit(trainset)
predictions = algo.test(testset)
accuracy.rmse(predictions)


# Filtrage collaboratif avec ajustement des moyennes
from surprise import KNNWithMeans
sim_options = {
    'name': 'cosine',
    'user_based': True
}
algo = KNNWithMeans(sim_options=sim_options)
algo.fit(trainset)
predictions = algo.test(testset)
accuracy.rmse(predictions)

# Filtrage collaboratif basé sur les utilisateurs
from surprise import KNNBasic

sim_options = {
    'name': 'cosine',
    'user_based': True
}

algo = KNNBasic(sim_options=sim_options)
algo.fit(trainset)
predictions = algo.test(testset)
accuracy.rmse(predictions)


#Filtrage collaboratif basé sur les modèle: Factorisation de matrices
from surprise import SVDpp

algo = SVDpp()
algo.fit(trainset)
predictions = algo.test(testset)
accuracy.rmse(predictions)


# Filtrage collaboratif basé sur les modèles Factorisation de matrices (matrice users × items)
algo_svd = SVD()
algo.fit(trainset)

# Prédictions sur test set
predictions = algo.test(testset)

# RMSE
rmse = accuracy.rmse(predictions)
print("RMSE sur le test set:", rmse)

## Sauvegarder les modeles et les encoders 

# Notebook ou script
import sys
from pathlib import Path

# Ajouter la racine du projet
project_root = Path.cwd().parents[1]
sys.path.append(str(project_root))

from src.load_save_data import save_to_both

# ================= SAUVEGARDE DES MODÈLES =================

save_to_both(algo, "knn_model.pkl")
save_to_both(algo_svd, "svd_model.pkl")

save_to_both(user_encoder, "user_encoder.pkl")
save_to_both(movie_encoder, "movie_encoder.pkl")

print("✅ Modèles et encodeurs sauvegardés (local + GCS)")
